Fix compilation on 32 bits platforms.
Only two OpenGL 4.x enums are problematic: GL_ALL_BARRIER_BITS and
GL_ALL_SHADER_BITS which should be 0xFFFFFFFF. We just represent
them by 0x7FFFFFFF and add a warning in the documentation of
the functions that may use them.

On these platforms the client can either OR all the other cases to get
a functional equivalent or try with that enumerant and pray that the
implementation doesn't throw a GL_INVALID_ENUM error.